Тесты разделены по пакетам, добавлен enum для кросс-браузерного тестирования. Исправлены ошибки с использованием аннотации @After для закрытия браузера
