+++ src/test/java/browserSwitch/browserList.java
@@ -1,0 +1,6 @@
+package browserSwitch;
+
+public enum browserList {
+    FIREFOX,
+    CHROME
+}
