+++ src/test/java/listOfAnswersTest/listOfAnswers.java
@@ -1,3 +1,5 @@
+package listOfAnswersTest;
+
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
