+++ src/test/java/orderFormTest/orderDetails.java
@@ -1,3 +1,5 @@
+package orderFormTest;
+
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
